backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0
asyncpg>=0.29.0
python-multipart>=0.0.6
passlib[bcrypt]>=1.7.4
bcrypt<4.0.0
pyjwt>=2.8.0
python-docx>=1.1.0
pypdf>=4.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
faiss-cpu>=1.7.4
openai>=1.0.0

